Rename notifications.rs to replica.rs
