+++ crates/store/src/state/replica.rs
